Change media rating from int to double

+++ backend/database/scripts/003-create_media_table.sql
@@ -4,7 +4,7 @@ CREATE TABLE media (
     title TEXT NOT NULL,
     description TEXT,
     cover_url TEXT,
-    rating INTEGER,
+    rating DOUBLE PRECISION,
     release_date DATE,
     created_at TIMESTAMP NOT NULL DEFAULT NOW(),
     updated_at TIMESTAMP NOT NULL DEFAULT NOW()
